enhancement: Refactored dependency injection extension methods to their appropriate layers
