+++ InventoryX.Infrastructure/DependencyInjection.cs
@@ -1,0 +1,21 @@
+using InventoryX.Application.Repository;
+using InventoryX.Infrastructure.Persistence;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace InventoryX.Infrastructure;
+
+public static class DependencyInjection
+{
+    public static IServiceCollection AddInfrastructure(this IServiceCollection services, IConfiguration configuration)
+    {
+        services.AddDbContext<AppDbContext>(options =>
+        {
+            options.UseSqlServer(configuration.GetConnectionString("DefaultConnection"));
+        });
+        services.AddScoped(typeof(IBaseRepository<>), typeof(BaseRepository<>));
+        services.AddScoped(typeof(ISalePurchaseRepository<>), typeof(SalePurchaseRepository<>));
+        return services;
+    }
+}
